+++ apps/native-component-list/src/navigation/ExpoApisStackNavigator.tsx
@@ -109,6 +109,13 @@ export const ScreensList: ScreenConfig[] = [
   },
   {
     getComponent() {
+      return optionalRequire(() => require('../screens/AppMetricsScreen'));
+    },
+    name: 'AppMetrics',
+    options: { title: 'App Metrics' },
+  },
+  {
+    getComponent() {
       return optionalRequire(() => require('../screens/AsyncStorageScreen'));
     },
     name: 'AsyncStorage',