+++ packages/expo-widgets/src/Widgets.types.ts
@@ -281,12 +281,12 @@ export declare class NativeWidgetObject extends SharedObject {
 
 export declare class NativeLiveActivityFactory extends SharedObject {
   constructor(name: string, layout: string);
-  start(props: string, url?: string): NativeLiveActivity;
+  start(props?: string, url?: string): NativeLiveActivity;
   getInstances(): NativeLiveActivity[];
 }
 
 export declare class NativeLiveActivity extends SharedObject<LiveActivityEvents> {
-  update(props: string): Promise<void>;
+  update(props?: string): Promise<void>;
   end(
     dismissalPolicy?: string,
     afterDate?: number,