+++ packages/expo-widgets/src/ExpoWidgets.ts
@@ -21,7 +21,7 @@ class WidgetStub {
 }
 
 class LiveActivityStub {
-  async update(_props: string): Promise<void> {}
+  async update(_props?: string): Promise<void> {}
   async end(
     _dismissalPolicy?: string,
     _afterDate?: number,
@@ -39,7 +39,7 @@ class LiveActivityStub {
 class LiveActivityFactoryStub {
   // eslint-disable-next-line @typescript-eslint/no-useless-constructor
   constructor(_name: string, _layout: string) {}
-  start(_props: string, _url?: string): NativeLiveActivity {
+  start(_props?: string, _url?: string): NativeLiveActivity {
     return new LiveActivityStub() as NativeLiveActivity;
   }
   getInstances(): NativeLiveActivity[] {