+++ packages/expo-widgets/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### 💡 Others
 
-- Temporarily make the Android config plugin opt-in with `enableAndroid`.
+- Temporarily make the Android config plugin opt-in with `enableAndroid`. ([#46463](https://github.com/expo/expo/pull/46463) by [@jakex7](https://github.com/jakex7))
 
 ## 56.0.15 — 2026-05-26